20170809 upstream rebase
* README: Make customizations in ~/dotfiles-local
* Remove git churn command
I added this four years ago but haven't been using it.
Not common enough to be included in company-wide dotfiles?
Can move to personal dotfiles if desired.
* Remove gvimrc
This originally came from jferris/dotfiles,
the parent of thoughtbot/dotifles.

I'm not a Vim GUI user.
Are others at thoughtbot also shell Vim users?

* Unbind the shortcut for suspend-client

Why:

I occasionally will try to zoom in on a pane with the key combination of
C-s then z. However, instead of having the agility to remove my finger
from the ctrl key, I clumsily continue to have it depressed while
applying pressure to the z key. This results in the `suspend-client`
command being executed by tmux. I am then left not knowing what to do
and I can't reattach to the session the that terminal window. It's very
disconcerting and I wish it hadn't happened.

This PR:

Removes the binding for `suspend-client`.

* Allow user to ignore some default vim plugins

The `UnPlug` command allows a user to remove a plugin from vim-plug's
list of plugins to install/activate. This command can be used to
selectively ignore shared plugins from `vimrc.bundles` without needing
to overwrite the entire file with a personal replacement.

This change is dependent on vim-plug continuing to use the `g:plugs`
global variable.

* Run `reset` in post-up hook to fix terminal issue

Some tmux users have reported that when running `rcup` that during the
post-up hook Vim would issue errors about input not coming from a
terminal and then the terminal would look and behave unexpectedly.

Running `reset` after the Vim plugin install step in `post-up` appears
to resolve the issue.

Fix #516.

* Add JavaScript syntax highlighting

It seems like JavaScript/ES6 is becoming more common for us to run into
in the wild compared to CoffeeScript
with the advent of Babel / transpiling.

Eventually, we'll probably want to remove kchmck/vim-coffee-script.
Users working on CoffeeScript applications would need to move that
plugin into their personal dotfiles.

* Don't rely on Bash in post-up hook

In 16219a31db5ca1f8c766d3c6066399950ff7f2c3 we implemented a post-up
hook that checked for a problematic `/etc/zshenv` file. OSX/macOS has
changed the file responsible for executing `path_helper` and loading the
system paths multiple times, and this script change issued a warning to
users if `path_helper` might be called in such a way that their paths
could be unexpectedly reordered.

The script used colored output to highlight this problem for the user,
and relied on Bash's script path introspection (through `$BASH_SOURCE`)
to print the location of the `post-up` hook to assist the user in
troubleshooting and understanding where our warning came from.
This change removes the hook's reliance on GNU Bash in favor of `sh`. It
uses the POSIX-compliant `$0` to determine the hook location (this
method is less robust than Bash's `$BASH_SOURCE`, but is available in
all POSIX shells) and removes colored output in the warning to be
compatible with shells that don't support colored output.
Fix #517.
* Skip compiled wordcode files (ie. zwc extensions)
* Combine conditional statements
* Revert "Combine conditional statements"
This reverts commit cae0bd04ea7472251c2ab1ce6ab1f7064d94e1f2.
This implementation appears to break sourced configurations. Reverting
until we figure out a resolution.
* Revert "Skip compiled wordcode files (ie. zwc extensions)"
This reverts commit 5178a8a1a83231c9650afc71d7acc69e53362973.
* Capitalize l in leaders
* Replace Syntastic with ALE
ALE is an Asynchronous Linting Engine which works with Vim 8 and NeoVim
to lint your code "as you type". It does so without blocking the main
thread by using the asynchronous features of both of those Vim versions.
A number of us in Boston have been using ALE for some time now with
great success. The asynchronous nature means linting does not block
other interactions with Vim and it's been nice to get linting feedback
more often than buffer write.
Many of us find the "as-you-type" linting to be performant enough to
actually be a distraction, so this PR tones that down some. With this
configuration, ALE will lint on cursor hold durations of 1 second and
when leaving insert mode in addition to buffer write, buffer read, and
file type change.
* added my own dotfiles to the repo
* Updated gitconfig
* Fixing styling and getting the zsh info first
* a little more styling
* Adding the upstream merge to the readme
* Updated README to link to the dotfiles-local repo
* Update README.md
* idea .idea and .iml to the git ginore and well as bundler and removed the random order from the rspec config
* Updated ignore
* Added mslav's simple tmux session saving script
* Modified the permissions for tmux session
* restore tmux windows layout on restore
* Added some keychain credential helpers for Sierra
* Remove tmux-session
* Add back random ordering to the default rspec config
* Move perl path and sqlite path changes from dotfiles zshrc to my zshrc.local